Move CAP UI implementations into a dedicated folder
